feat: add test, remove sandbox
