+++ validate/isbn/package.json
@@ -50,13 +50,8 @@
     "@flatfile/listener": "^1.1.0"
   },
   "devDependencies": {
-    "@flatfile/hooks": "^1.5.0",
     "@flatfile/rollup-config": "^0.1.1",
-    "@types/jest": "^29.5.13",
-    "jest": "^29.7.0",
-    "rollup": "^4.22.5",
-    "ts-jest": "^29.2.5",
-    "typescript": "^5.6.2"
+    "@flatfile/utils-testing": "^0.3.0"
   },
   "repository": {
     "type": "git",
